Show different color for NUM according to watermark
We need know the watermark for each NUMA when debugging memory
related issues, esp. too few memory for apps to work normally.
In this case, we have to manually calculate those watermarks to
compare if the 'free' memory is lower than min or low or high
watermark, which is quite time-consuming. Let's optimize this
by changing NUM color according to the comparation:
 - free_mem <= min_watermark: red
 - min_watermark < free_mem <= low_watermark: cyan
 - low_watermark < free_mem <= high_watermark: green

As these watermarks are hardly updated once reboots, for
performance consideration, we employ to check if the value from
/proc/fs/vm/watermark_scale_factor changes, instead of frequently
reading /proc/zoneinfo file during each interval. If the
watermark_scale_factor is not changed, continue to use the previous
value, or else read /proc/zoneinfo again to get the latest value.

Reported-by: Gang Li <ligang.bdlg@bytedance.com>
Signed-off-by: Fei Li <lifei.shirley@bytedance.com>